// 윤-정윤주/4주차/chapter4-1/src/types/Credits.ts
export interface Cast {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string | null;
    cast_id: number;
    character: string;
    credit_id: string;
    order: number;
}

export interface Credits {
    id: number;
    cast: Cast[];
}